requirements.txt
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.0.3
Flask-SocketIO==5.3.7
mtcnn==0.1.1
mysql-connector-python==9.0.0
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
requests==2.32.3